TST: split tests/format/test_format.py
closes #15531
